move cffi code into windowsOS.api.py
